Move MySQL state tests to pytest
