+++ tests/filename_map.yml
@@ -57,9 +57,9 @@ salt/modules/influxdb08mod.py:
   - pytests.unit.states.test_influxdb08_user
 
 salt/modules/mysql.py:
-  - unit.states.test_mysql_user
-  - unit.states.test_mysql_query
-  - unit.states.test_mysql_grants
+  - pytests.unit.states.test_mysql_user
+  - pytests.unit.states.test_mysql_query
+  - pytests.unit.states.test_mysql_grants
 
 salt/modules/openvswitch.py:
   - pytests.unit.states.test_openvswitch_port
